refactor: separate fetching logic from UI components
