+++ galasa-ui/src/components/test-runs/TestRunsTable.tsx
@@ -31,7 +31,7 @@ import StatusIndicator from "../common/StatusIndicator";
 import { useEffect, useMemo, useState } from "react";
 import { useRouter } from "next/navigation";
 import ErrorPage from "@/app/error/page";
-import { TestRunsData } from "@/app/test-runs/page";
+import { TestRunsData } from "@/lib/data/testRuns";
 import {MAX_RECORDS} from "@/utils/constants/common";
 import { useTranslations } from "next-intl";
 import { InlineNotification } from "@carbon/react";
@@ -147,10 +147,10 @@ export default function TestRunsTable({runsListPromise}: {runsListPromise: Promi
   // Generate the time frame text based on the runs data
   const timeFrameText = useMemo(() => {
     if (!rawRuns || rawRuns.length === 0) {
-      return translations("noData");
+      return translations("noTestRunsFound");
     }
 
-    let text = translations("timeFrameText.default");;
+    let text = translations("timeFrameText.default");
     const dates = rawRuns.map((run) =>
       new Date(run.testStructure?.queued || 0).getTime(),
     );
